requirements.txt
pyyaml
enquiries
requests
tabulate
jmespath
argparse
